DDT: fix off-by-one error in lower bound calculation
std::min_element(first, last) returns a pointer to the minimum value of
an array between first (inclusive) and last (exclusive). However, count
already points to the first element *after* the array, such that count+1
includes uninitialized memory in the search for the minimum.
Uninitialized memory tends to have zero values, leading to crashes
(or accesses to random memory locations) when accessing lb.

Change-Id: I306c13370e18ac9865af844bfc25d3e973705510