Bug #1616: Don't exit cputopology until reducer has been applied
For any given process, the worker threads that are supposed to reduce
and the comm thread won't exit cputopology until the reducer for
the second CmiReduce has been executed.

This guarantees that sending of messages corresponding to the CmiReduce
will happen even if worker threads eventually encounter a node barrier
after cputopology (since worker threads won't be needed anymore to
advance the reduction).

Change-Id: Id6d130e82fda245308f82205b00f5b5cf1988bc4