Fix a potential race condition in examples/rdma/simpleRdma
The sdag test case in simpleRdma should wait for the completion of
the sending of the rdma parameters before progressing to the next
iteration and possibly load balancing. Added simple counters to ensure
that the next iteration begins only after completion of rdma sends and
after receiving all the when clause messages.

Change-Id: Ida35a486e2137637251b1f45bfdbf451bbc40fc9