Ensure function pointer types have correct parameter types
In C, declaring function prototypes with an empty parameter list means
the compiler will treat any combination of parameters and parameter
types as valid. In C++, this can only mean the function takes no
parameters, which can be more robustly declared in C using the keyword
"void" inside the parentheses.

Change-Id: I055873181eb42847b30d664cef39093c2a8cdff6