AMPI: add documentation about AMPI codes
This is a copy of
https://wiki.illinois.edu/wiki/display/ppl/Adaptive+MPI+%28AMPI%29+Codes

This web site is only available to people with a UIUC account. By adding
the list to the manual, other people can benefit (for performance
comparisons, for example).

Codes have been moved to our git repository:
git clone ssh://charm.cs.illinois.edu:9418/benchmarks/ampi-benchmarks

Change-Id: Ie195888f0c09fa1d3dc0e7c199bb85432a0cc1b9