Bug #1690: Modify /zerocopy/simpleZeroCopy/Makefile to run test with fewer chares
This bug caused the verbs-smp crash in autobuild. The crash occurred because of the
default send tokens in verbs being insufficient for the previous test argument i.e 200.
For this reason, the test argument was changed from 200 chares to 32 chares, which works
with the default number of send tokens (1000 send tokens).

Change-Id: I13a75b9a15c62d432af5bb05132fd524379e1be1