Revert the test change for malloc_info
This change reverts the test for malloc_info as it fails
on some platforms. Upon a closer look the failure has to do
with malloc(size) not being recorded by ps memory reporter. And
on platforms where malloc_info is falling through to ps reporter,
the correct size is not recorded.

CmiAlloc(size) on the other hand is indeed recorded by ps, so
reverting this test will address the build failure for now. Will add
proper test case for malloc_info supported platforms later.

Change-Id: Iafb9e27e365aafeb5f78e6d77c06a17607d58cff