Shrink size of Charm message envelope from 80 to 64 Bytes
This is possible because CmiReservedHeaderSize is 28, so the first
28 Bytes are reserved for Converse. ck::impl::u_type requires 8-Byte
alignment, so there were 4 free Bytes between the Converse header and it.
By moving 'pe' to occupy that space, we bring the size of envelope down
to a naturally 8-Byte aligned size, so the 'align' field becomes a zero
length char array, rather than using 8 Bytes. In total, envelope used to
be 75 Bytes, which the compiler padded to 80 Bytes. With 4 fewer bytes
due to the redordering and 8 fewer bytes due to alignment, it is now 64
bytes. This is for a build of Charm --with-production.

Change-Id: I73f6c41ba79a6695d9292023516d4aea8bfd00ab