Automate process of building libcharm.so for charmpy
Changes in charmc by Evan Ramos.

Added `libcharm` and `charmpy` rules to Makefile to generate
libcharm.so automatically.

Added checks in configure.ac to find the correct compiler/linker
flags to build the library.

For Charmpy, the build line to generate the library is like this:
./build charmpy netlrts-linux-x86_64

If the target is `charmpy`, the build script will automatically add
--build-shared (on non-Windows environment) and --enable-charmpy

Change-Id: I62894f3a18fa54625170f21df07837df6efaccfe