Bug #1871: GNI RDMA - Direct RGET/RPUT calls through the comm thread
With this fix, RGET/RPUT calls are performed by the comm thread on
behalf of the worker thread. Previously, this was done by the worker
threads themselves and this was sometimes causing a segmentation violation
during the GNI_GetCompleted() method as a shared completion
queue was being unsafely accessed by both the comm thread and the worker
threads.

This bug fix also resolves the gni-crayxc-smp autobuild crash seen during
the execution of get_put_pingpong.

Change-Id: I965f39164d8548b02402f21f884276b5ad5198e0