Refactor the Zerocopy Direct API
Use a single dynamically allocated object of the type NcpyOperationInfo,
as a single argument, instead of passing many parameters. This design leads
to fewer smaller allocations in the machine layer code and allows to simplify
code significantly. It also makes the code more maintainable and flexible
towards changing behavior or passing new information down to the Lrts RDMA methods.

This commit also adds the feature (Feature #1919) for an optional arbitrary pointer
being passed by the user before the zerocopy operation (get/put). This reference
pointer is returned back to the user in the callback using the CkNcpyAck object.

Change-Id: I0ec2320b664aec8dee6d0a4e327c0ee689326e4f