Bug #1932: Fix race condition in converse pingpong examples
The bug was caused by a race condition between the two PEs.
Because of a call to CmiInitCPUTopology, the scheduler is polled
for messages inside the slower PE. (CsdSchedulePoll() called inside
CmiInitCPUTopolgy in cputopology.c) and this causes the slower PE
to receive a message from the faster PE before it registers its
message handlers (which happens in the pingpong code after the
topology initialization). This is caused by the faster PE being ahead in
execution and sending a message with node1Handler before the slower PE even
registers the node1Handler. This race condition causes the bug, leading
to throwing an invalid handler error where the handler index of the
received message is greater than the maximum registered handler index.

The fix is to call CmiCPUTopologyInit after the initialization of the
handlers in order to avoid receiving messages with unregistered message
handlers.

Change-Id: I6770b8d246b6dee5a035946b2625bd15a5acfbbd