Bug #1929: Initialize cmitls segments by copying the current one
This fixes crashes in Fortran AMPI programs during formatted print and
write statements. Internally, libgfortran uses ctype.h functionality,
which depends on thread-local storage to account for per-thread locale
changes, and it was dereferencing a pointer that was left NULL due to
cmitls only copying the initial data found in the ELF header instead of
the full segment.

Change-Id: Iedd31e32e6a165ad05a292b1a3cc57dc570b59bc