Use refNum to determine ncpy transfer completion order
Previously, the completion order of the ncpy transfers
in examples/charm++/zerocopy/direct_api/unreg/simple_get
was assumed to be in the order of calling the get invocation.
This is fixed by setting reference numbers in the callbacks and using
those reference numbers to determine the get completion order.

Change-Id: I4e2c28cc5fc64a6fa1e7f629863bbf12e0e7ef07