Bug #1961: Convert LBDB's list of objects to use a linked list embedded in an array to prevent unbounded growth
Previously, the list of objects maintained in the LBDB grew without
bound, as nothing was ever deleted and new objects were pushed to the
back of the vector. Now we try to fill empty spots in the vector
before pushing to the back by embedding a linked list of the empty
spots into the vector.

Change-Id: I276ec24611719a82f941d16830508ffdf33c851b