netlrts: close charmrun socket on abort
On abort, the socket that sends messages to charmrun was not
explicitly closed before exit. On Windows, this has the effect
that abort messages are often not printed, and instead charmrun
exits with:

"Charmrun> error on request socket to node x '127.0.0.1'

Not sure if it's also necessary to wait for data to be sent
after closing the socket (maybe with SO_LINGER option), but my
testing on Windows suggests that this fix is enough.

Change-Id: I15489b1e39b34b159931312c17c8f195bc96d793