Fix bug in MPI machine layer related to the Ncpy API
The bug is related to linked list states when messages are sent
in the callback function invoked inside ReleasePostedMessages.
Previously, the end_sent link list was stale when a new message was
added to the linked list while iterating across the list inside
ReleasePostedMessages. With the fix, the link list state is updated
before invoking the acknowledgement function to handle sends while
iterating across the list.

Change-Id: Ib06b3ab7375c13f328df9502976b5fc900fbc03f