Feature #195: Delete example which uses older RDMA API: examples/charm++/onesided
The example illustrates the usage of the older RDMA API by
performing a Put or a Get operation, with or without callbacks.
This is already illustrated in the newer Zerocopy Direct API examples
which can be found in examples/charm++/zerocopy/direct_api. To avoid that
redundancy, I am in favor of deleting this example rather than
resurrecting it with the newer API.

Change-Id: I9528364f1486c284df0c5c66025c6827ce0f90e1