Feature #1742: Use ZC infrastructure for bcast of large readonly variables
A macro CMK_ONESIDED_RO_THRESHOLD is used to specify the threshold
which represents the size in bytes over which the Zerocopy API is used
to broadcast the large readonly variables. Currently, it is defined
in src/arch/util/lrts-common.h. In the future, it has to be a layer
dependent macro that is computed based on experiments and defined in
src/arch/<layer>/conv-common.h.

Change-Id: I261345f2e8bc1520c6b67f08ae8abf79f9534ce3