New thread CmiNodeBarrierCount implementation for win-smp
The new barrier implementation uses 2 separate barriers: one for
CmiNodeBarrier, and a different one for CmiNodeAllBarrier.

Trying to reuse the same barrier with different thread count leads
to incorrect behavior under some race conditions. On Windows, this
is observed in some cases during charm initialization, where the
commthread in the last CmiNodeAllBarrier gets stuck because of 3
CmiNodeBarriers that are called in initDone.

Change-Id: Idcfd7a147b60b7b9bdba548efafd45ff5ef746f4