charmc: Don't skip linking if user explicitly specified an output file
This allows generating dependencies while compiling

If you have a directory layout like

src/
   foo/
      a.cpp
   a.cpp

this changeset fixes a seemingly unrelated bug where compiling
'foo/a.cpp' from 'src' using
'charmc -MMD -MP -MF foo/a.d -c foo/a.cpp -o foo/a.o' produces
'src/a.o' instead of 'src/foo/a.o'- overwriting (or will be
overwritten by) the object generated by 'src/a.cpp'. This happens
because of the test 'if [ ! -z "$SKIPLINK" ]' on line 1854
(in this commit; 1848 in HEAD).

A simpler fix would be to not set SKIPLINK when a dependency flag is
read in 'processArgs'. However, that would change the meaning when
the user does not specify an output file (and linking should be
skipped) and wants to generate dependency files.

This bug likely never showed up because it works fine when the
source being compiled is in the same directory from which charmc is
invoked as most compiler's default output location is the pwd.

Change-Id: I8b28e64ba46f060039bd3c4f0e88b8d842ddb3bc