AMPI: shrink size of gather(v) and non-commutative redn msgs for comm sizes <64K
For communicators with less than or equal to 64K ranks, we can fit the
rank of the contributor into an unsigned short int rather than an int
in the concat reduction msg. This applies to all (I)(All)Gather(v)
calls and all non-commutative reductions.

At its most effective (for a communicator with size 64K), this reduces
the final CkReductionMsg's size by 128KB.

Change-Id: I58024f14bcf861c20e159eea366bfd2ec532d90e