Clean up C linkage specific to the C++ migration
These are the principles applied in this patch:

* Remove most instances of CMI_EXTERNC, since they were private to the
RTS and we no longer have a mix of C and C++ files.
* Replace remaining instances of CMI_EXTERNC with CLINKAGE
* Remove instances of extern "C" { } surrounding #includes
* Remove spurious instances of extern "C" at definition sites
* Ensure CLINKAGE/extern "C" is on the same line as what follows it,
whether it is a curly brace or a declaration, for improved greppability

extern "C" should always be attached to a declaration in a header in
order to avoid mismatches between definitions and callers elsewhere.

Change-Id: I4c6ef8351d53636e98cbc02e25048a851f05b790