Do not setObjPtr for nodegroup messages
Previously, on receiving nodegroup messages, 'setObjPtr' was called
to set the local object pointer in the received message. This was later
used to invoke the entry method on the local object (by calling
'getObjPtr'). Calling 'setObjPtr' sets the ptr field in the union
variable called 'type', overwriting the CkGroupID field of the union.

For supporting the Zerocopy Post API for nodegroups, this CkGroupID field
was necessary to be retained correctly in the envelope. For this reason,
this patch does not call 'setObjPtr' and instead directly passes the local obj
ptr to 'invokeEntry' method without having the need to set/get the local
obj pointer in the envelope. Additionally, nodegroup messages do not
have their type set to 'ForChareMsg' anymore and the function
'_processForChareMsg' is removed from the code as it is no longer used.

Change-Id: Id2eb75e9f08719d8c9e2f42261abf711e25694c4