Fix an issue where turning off manual LB breaks AtSync
The problem previously was that when manual LB is on, all check-ins to the
barrier (including from the batsyncer) were entirely ignored. So if then manual
LB was turned off, those calls were lost forever and the barrier would never be
satisfied. This change preserves the calls to the barrier, but keeps the barrier
off until manual LB is turned off.

Change-Id: I0738abcf07af4bc922d73cda40a4c8bf3ddd676b