ZC RO Bcast: Initialize _numPendingRORdmaTransfers with numZerocopyROops
Previously, _numPendingRORdmaTransfers was initialized to 0 and
incremented inside the readonlyGet method, which was called when
the RO variable was being unpacked. This caused a race condition
where the the first RDMA Get operation could be completed before
_numPendingRORdmaTransfers was incremented again for a second RO
variable. Hence, this fix initializes _numPendingRORdmaTransfers
before launching any RDMA operations to avoid completion of an
RDMA operation before _numPendingRORdmaTransfers is initialized
to the total number of RDMA RO variables.

Change-Id: Ie53ee802c3eabfbb2d0c3cf57776a9365fa050a7