ZC API: Do not block CK_BUFFER_PREREG from de-registering in deregisterMem
At the generic layer (ckrdma.h), CK_BUFFER_PREREG buffers should be
allowed to de-register as they are implementation dependent. For example,
for GNI and Verbs, which support CK_BUFFER_PREREG, the de-registration
is blocked at the LRTS implementation. On the other hand, layers like
OFI which do not support CK_BUFFER_PREREG should be allowed to de-register
that buffer in the LRTS method.

Change-Id: I145277dfa3196658920b2c164df53a9865cf2464