Fix a race in MetaLB regarding 0 object PEs (#2360)
If the intial periodic call to check for 0 object PEs triggered at a weird
time (ie during migration), it could incorrectly detect an empty PE and cause
a mismatched reduction contribution. This makes it so the periodic call is only
checked if we have not received any load contributions at the time it triggers.